refactor(rust): resolver warnings triviales de clippy
131 → 30 warnings. Los 30 restantes son too_many_arguments,
que requieren refactor real.

- clippy --fix automático: explicit_auto_deref, casts redundantes,
  div_ceil manual, map_or, closures redundantes, RangeInclusive, etc.
- &PathBuf → &Path en config, runner, cloud y micromamba
- clamp() en pdf_commands y ort_runner
- needless_range_loop en pixel_commands y csv_import
- format! anidado en colab y vertex_custom: gcs_results a binding
- unnecessary_unwrap en distribution: is_some()+unwrap → if let Some
- ort_runner: doc comment huérfano de un run_inference ya borrado
  estaba pegado a preprocess_image
- tests.rs: #![cfg(test)] duplicado (lib.rs ya lo aplica)

+++ src-tauri/src/export/unet_masks.rs
@@ -49,8 +49,8 @@ pub fn export<F: Fn(f64)>(
 }
 
 fn generate_mask(image: &ImageEntry, project: &ProjectFile) -> Result<Option<Vec<u8>>, String> {
-    let w = image.width as u32;
-    let h = image.height as u32;
+    let w = image.width;
+    let h = image.height;
 
     // Create grayscale mask (background = 0)
     let mut mask_img = GrayImage::from_pixel(w, h, Luma([0u8]));
